requirements.txt
python-dotenv==1.0.0
openai==1.12.0
pandas==2.2.0
python-dateutil==2.8.2
mcp-server-openbnb==1.0.0
gspread==5.12.3
oauth2client==4.1.3 